Extends substitute management to owner and team managers

+++ app/Http/Controllers/UserSubstituteController.php
@@ -3,8 +3,6 @@
 namespace App\Http\Controllers;
 
 use App\User;
-use App\UserSubstitute;
-use Exception;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 
@@ -15,28 +13,20 @@ class UserSubstituteController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function index()
+    public function index(Request $request)
     {
-        $candidateSubstitutes = \App\Team::find(session('currentProject'))
+        $substitutableUsers = \App\Team::find(session('currentProject'))
             ->subject_managers()
-            ->where('id', '!=', auth()->user()->id)
-            ->where('site_id', auth()->user()->ProjectSite)
+            ->where(function ($query) use ($request) {
+                if (!(Auth::user()->isAbleTo('manage-teams', $request->currentProject->team->name) or Auth::user()->owns($request->currentProject, 'owner'))) {
+                    $query->where('site_id', Auth::user()->project_site);
+                }
+            })
+            ->with('currentSite')
+            ->orderBy('site_id')
+            ->orderBy('firstname')
             ->get();
-
-        // $currentSubstitute = auth()->user()->substitute()->get();
-        $currentSubstitute = Auth::user()->substitute->first();
-        
-        return view('userSubstitutes.index', compact('currentSubstitute', 'candidateSubstitutes'));
-    }
-
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function create()
-    {
-        //
+        return view('userSubstitutes.index', compact('substitutableUsers'));
     }
 
     /**
@@ -47,9 +37,16 @@ class UserSubstituteController extends Controller
      */
     public function store(Request $request)
     {
+        $validatedData = $request->validate([
+            'user_id' => 'required|exists:users,id',
+            'substitute_id' => 'required|exists:users,id'
+        ]);
+        if (!$this->checkPermissions(intval($validatedData['user_id']), $request->currentProject)) {
+            return back()->with('error', 'You do not have permission to edit that user');
+        }
         try {
-            $user = auth()->user();
-            $user->substitute()->sync([$request->substitute_id => ['team_id' => session('currentProject')]]);
+            $user = User::find($validatedData['user_id']);
+            $user->substitute()->sync([$validatedData['substitute_id'] => ['team_id' => session('currentProject')]]);
         } catch (\Throwable $th) {
             return back()->with('error', 'Substitution Failed: ' . $th->getMessage());
         }
@@ -62,32 +59,18 @@ class UserSubstituteController extends Controller
      * @param  \App\UserSubstitute  $userSubstitute
      * @return \Illuminate\Http\Response
      */
-    public function show(UserSubstitute $userSubstitute)
+    public function show(Request $request, User $user)
     {
-        //
-    }
-
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  \App\UserSubstitute  $userSubstitute
-     * @return \Illuminate\Http\Response
-     */
-    public function edit(UserSubstitute $userSubstitute)
-    {
-        //
-    }
-
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  \App\UserSubstitute  $userSubstitute
-     * @return \Illuminate\Http\Response
-     */
-    public function update(Request $request, UserSubstitute $userSubstitute)
-    {
-        //
+        if (!$this->checkPermissions($user->id, $request->currentProject)) {
+            return back()->with('error', 'You do not have permission to edit that user');
+        }
+        $candidateSubstitutes = \App\Team::find(session('currentProject'))
+            ->subject_managers()
+            ->where('id', '!=', $user->id)
+            ->where('site_id', $user->project_site)
+            ->get();
+        $currentSubstitute = $user->substitute->first();
+        return view('userSubstitutes.show', compact('user', 'currentSubstitute', 'candidateSubstitutes'));
     }
 
     /**
@@ -98,13 +81,26 @@ class UserSubstituteController extends Controller
      */
     public function destroy(Request $request)
     {
+        $validatedData = $request->validate([
+            'user_id' => 'required|exists:users,id'
+        ]);
+        if (!$this->checkPermissions(intval($validatedData['user_id']), $request->currentProject)) {
+            return back()->with('error', 'You do not have permission to edit that user');
+        }
         try {
-            $user = auth()->user();
-            // $user->substitute()->detach($request->substitute_id, ['team_id' => session('currentProject')]);
+            $user = User::find($validatedData['user_id']);
             $user->substitute()->detach();
         } catch (\Throwable $th) {
             return back()->with('error', 'Substitution Deactivation Failed: ' . $th->getMessage());
         }
         return back()->with('message', 'Substitution Deactivated');
+    }
+
+    private function checkPermissions(int $user_id, \app\project $currentProject)
+    {
+        if ($user_id === Auth::user()->id or Auth::user()->isAbleTo('manage-teams', $currentProject->team->name) or Auth::user()->owns($currentProject, 'owner')) {
+            return true;
+        }
+        return false;
     }
 }
